Refactored components to fetch project data from context
Instead of passing project data directly as prop to SingleProjHeader, ProjParagraphs, SProjSection and ProjectRelatedTechs components, it now fetches the data from the Utils context. Additionally, this update adds a loading spinner to the SProjSection and ProjParagraphs components to display while project data is being fetched. The ProjectRelatedTechs component was extended to include a heading for related technologies.